Refactor FilterBar to close over state
This refactor prevents two `<FilterBar />` components from clashing with
each other if more than one gets instantiated during the application
runtime.

Previously, since the list of filters was stored in the module's static
code area, more than one instance of the React component would directly
modify that state and the two `<FilterBar />`s would fight over that
single mutable state.

Now, each `<FilterBar />` is actually the result of creating a new
closure with the state self-contained. A drawback to this approach is
that it's not enough to simply import the filter bar component. Instead,
the factory is imported and a new `<FilterBar />` requires calling that
factory. On the plus-side, this further removes the inter-module
dependencies that previously existed in the shared-state model, leaving
components independent and isolated.

Props to @seanastephens for the suggestion