Abstract app to allow additional middleware
Previously, with the store being created in the app module at
initialization-time, there was a barrier to providing external
middleware or store enhancers because of the lack of passing that data
into the initialization code.

This was been rebuilt behind a factory to allow passing in custom
middleware and enhancers through the use of a non-default app factory
constructor.

By default, there will be no additional store enhancer other than what
is required to run the app. However, by using the custom constructor, we
can create a local development app and pass in additional tooling such
as a redux logger to record actions/state-transitions or even a routing
store enhancer to hook up URL mapping for props.