Add URL-based routing to local dev environment
Nominally the app routes based solely off of the props set for things
like the selected note and selected filter. This is great because it
keeps the app in a purer state as a React component than were it to
incorporate something like `express.js` routing or even `react-router`.

However, it's definitely useful to allow URL-based routing when
developing locally in order to make things like back-navigation work in
the browser. Sometimes it would be better also to simply start the app
in a specific state specified by the selected note, etc...

This patch refactors a few systems in order to allow coupling in a
router or other similar app enhancement in a local environment while
keeping all of that logic separate from the app proper.

 - Instead of exporting the App component by default, a new factory
   function has been created and the default export is now the component
   that results from calling that factory with the default parameters.

 - The new factory accepts an optional list of middleware/store
   enhancers that then get passed into `createStore()` which is closed
   over in that factory function itself. The factory returns both the
   app component _and_ the created store.

 - The routing wrapper creates a new app with special middleware to
   intercept actions which should trigger path changes and pushes those
   into the history state. It also sends out a special dispatch to the
   store whenever the URL parameters change. This special dispatch gets
   intercepted by the middleware and transforms into the appropriate
   action which would cause the corresponding change, but doesn't update
   the browser history state.

 - Composability all the way down.