Clean up some navigation issues with URL-based routing
This cleans up the way the app loads if loading from a URL other than
the root '/'. Now, it can load from a state where the URL specifies the
filter and or the note id. Additionally, if a specific notification is
selected on load, this will insert a new back state in the browser
history so that navigation back to the note list works as expected.