Add oAuth wrapper around local app
Previously to run the app, you needed to store a copy of your oAuth
authentication token in a `config.js` file so that the API calls would
work. This could be a tedious process if you didn't already have the
token.

Now, the app will automatically request authentication through an
implicit oAuth request and redirect on success. The app stores the token
in a cookie for persistence across sessions.