Load mystery-man SVG while waiting for avatars to load
Previously, nothing was showing in the spot for the avatar in the list
view while the images were loading from the Gravatar service.

Now, an SVG version of the mystery-man loads and is automatically
replaced once the image data loads in the background.

This is particularly noticeable on slow or laggy internet connections.