Add `hasReplied` field to notifications
This patch adds the `hasReplied` property to a notification, indicating
that the notification embodies both a comment and that the user has
replied to that comment.

The app indicates that a comment has been replied to in the list view by
prefixing the subject line with a reply arrow graphic.