Add block parsing for text ranges
This commit adds basic support for parsing the text ranges in the
notification blocks. This essentially means that we are introducing
styling on the note content, though that styling is severaly limited at
this point.

The algorithm for building the formatting uses a form of a topological
sort of the text ranges, then recursively breaks down the remaining text
in the string for each interval. This builds an object specifying
special properties of the range and which nests according to the ranges.

```js
const block = {
	text: 'Math is fun!',
	ranges: [
		{ indices: [ 0, 11 ], type: 'blockquote' },
		{ indices: [ 0, 4 ], type: 'b' }
	]
}

// --> turns into...
const tree = [
	{
		type: 'blockquote',
		children: [
			{
				type: 'b',
				children: [ 'Math' ]
			},
			' is fun!'
		]
	}
]
```

The view layer can take this plain object and translate it according to
the goal: the app currently turns this into a React component hierarchy.

It's worth noting that ranges that aren't understood by the app default
to a textblock that gets surrounded by a span with a CSS class
corresponding to the `type` of the node.