Fixed bug in Avatar
Previously, if the avatar unmounted while it was still loading its
image, it would attempt to `setState()` without being mounted.

Now, instead of using `avatar.onload = ...` we are using the
`addEventListener()` method to hook into the `load` event and
correspondingly calling `removeEventListener()` from
`componentWillUnmount()` to prevent that erroneous `setState()` from
being called.

This _did_ require setting `this.avatar` in the component but as it
should be the only instance per component, this shouldn't give us any
trouble.