Swap `oAuthToken` prop dependency for `wpcom` dependency
Previously, the app required an `oAuthToken` so that it could create an
authenticated connection to the WordPress API.

This patch changes the behavior so that it's the actual connection that
gets passed down as a prop. This is to prepare the app for a possible
integration into the Automattic/wp-calypso project.